adk-native/requirements.txt
# ADK-Native Vacation Planner Requirements
# Pure Google ADK implementation

# Google Agent Development Kit
google-genai>=0.3.0
google-adk>=0.1.0

# HTTP clients for API calls
httpx>=0.27.0

# Logging
loguru>=0.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Optional: For enhanced features
python-dotenv>=1.0.0  # Load env vars from .env file
